mpegtspacketizer: Store PCR time/byte offsets in 64 bit integers
32 bit integers are going to overflow, especially the PCR offset to
the first PCR will overflow after about 159 seconds. This makes playback
of streams stop at 159 seconds as suddenly the timestamps are starting
again from 0. Now we have a few more years time until it happens again
and 64 bits are too small.

+++ gst/mpegtsdemux/mpegtspacketizer.h
@@ -99,10 +99,10 @@ typedef struct
 typedef struct _PCROffset
 {
   /* PCR value (units: 1/27MHz) */
-  guint32 pcr;
+  guint64 pcr;
 
   /* The offset (units: bytes) */
-  guint32 offset;
+  guint64 offset;
 } PCROffset;
 
 /* Flags used on groups */
